+++ examples/bootstrap_policies/main.tf
@@ -14,3 +14,12 @@ module "boostrap_full" {
   allow_eks_management = true
   allow_iam_management = true
 }
+
+module "boostrap_full_role" {
+  source      = "../../modules/bootstrap_role"
+  policy_name = "streamnative-mc-bootstrap-full"
+  role_name   = "streamnative-role"
+
+  allow_eks_management = true
+  allow_iam_management = true
+}